Fix package structure of unit tests.
Fixes #4
