+++ test/edu/wright/cs/raiderplanner/model/VersionControlEntityTest.java
@@ -19,7 +19,7 @@
  *
  */
 
-package model;
+package edu.wright.cs.raiderplanner.model;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
